final booking refactor
